Fix init functions not being called
The onInit.ts function was being bundled into index.js, so it was not
being found by the addon events.json. As a side efect, because
events.json was not valid, the onInit.mts was not being called either.
I fixed it by moving all backend logic into onInit instead of index.